requirements.txt
absl-py                      2.3.1
anyio                        4.9.0
argon2-cffi                  25.1.0
argon2-cffi-bindings         21.2.0
arrow                        1.3.0
asttokens                    3.0.0
astunparse                   1.6.3
async-lru                    2.0.5
attrs                        25.3.0
babel                        2.17.0
beautifulsoup4               4.13.4
bleach                       6.2.0
certifi                      2025.7.14
cffi                         1.17.1
charset-normalizer           3.4.2
colorama                     0.4.6
comm                         0.2.2
contourpy                    1.3.2
cycler                       0.12.1
debugpy                      1.8.15
decorator                    5.2.1
defusedxml                   0.7.1
executing                    2.2.0
fastjsonschema               2.21.1
flatbuffers                  25.2.10
fonttools                    4.59.0
fqdn                         1.5.1
gast                         0.6.0
google-pasta                 0.2.0
grpcio                       1.74.0
h11                          0.16.0
h5py                         3.14.0
httpcore                     1.0.9
httpx                        0.28.1
idna                         3.10
ipykernel                    6.29.5
ipython                      9.4.0
ipython_pygments_lexers      1.1.1
isoduration                  20.11.0
jax                          0.6.2
jaxlib                       0.6.2
jedi                         0.19.2
Jinja2                       3.1.6
json5                        0.12.0
jsonpointer                  3.0.0
jsonschema                   4.24.0
jsonschema-specifications    2025.4.1
jupyter_client               8.6.3
jupyter_core                 5.8.1
jupyter-events               0.12.0
jupyter-lsp                  2.2.5
jupyter_server               2.16.0
jupyter_server_terminals     0.5.3
jupyterlab                   4.4.4
jupyterlab_pygments          0.3.0
jupyterlab_server            2.27.3
kiwisolver                   1.4.8
libclang                     18.1.1
Markdown                     3.8.2
markdown-it-py               3.0.0
MarkupSafe                   3.0.2
matplotlib                   3.10.3
matplotlib-inline            0.1.7
mdurl                        0.1.2
mediapipe                    0.10.21
mistune                      3.1.3
ml_dtypes                    0.5.1
mpmath                       1.3.0
namex                        0.1.0
nbclient                     0.10.2
nbconvert                    7.16.6
nbformat                     5.10.4
nest-asyncio                 1.6.0
notebook                     7.4.4
notebook_shim                0.2.4
numpy                        1.24.3
oauthlib                     3.3.1
opencv-contrib-python        4.11.0.86
opt_einsum                   3.4.0
optree                       0.17.0
overrides                    7.7.0
packaging                    25.0
pandocfilters                1.5.1
parso                        0.8.4
Pillow                       10.0.1
pip                          24.0
platformdirs                 4.3.8
prometheus_client            0.22.1
prompt_toolkit               3.0.51
protobuf                     3.20.3
psutil                       7.0.0
pure_eval                    0.2.3
pyasn1                       0.6.1
pycparser                    2.22
pydantic_core                2.14.1
Pygments                     2.19.2
pyparsing                    3.2.3
python-dateutil              2.9.0.post0
python-dotenv                1.1.1
python-json-logger           3.3.0
python-multipart             0.0.6
pywin32                      311
pywinpty                     2.0.15
PyYAML                       6.0.2
pyzmq                        27.0.0
referencing                  0.36.2
requests                     2.32.4
rfc3339-validator            0.1.4
rfc3986-validator            0.1.1
rich                         14.1.0
rpds-py                      0.26.0
scipy                        1.16.0
Send2Trash                   1.8.3
sentencepiece                0.2.0
setuptools                   65.5.0
six                          1.17.0
sniffio                      1.3.1
sounddevice                  0.5.2
soupsieve                    2.7
stack-data                   0.6.3
sympy                        1.14.0
tensorboard                  2.19.0
tensorboard-data-server      0.7.2
tensorflow                   2.19.0
tensorflow-estimator         2.14.0
tensorflow-io-gcs-filesystem 0.31.0
termcolor                    3.1.0
terminado                    0.18.1
threadpoolctl                3.6.0
tinycss2                     1.4.0
tornado                      6.5.1
tqdm                         4.67.1
traitlets                    5.14.3
types-python-dateutil        2.9.0.20250708
typing_extensions            4.14.1
uri-template                 1.3.0
urllib3                      2.5.0
wcwidth                      0.2.13
webcolors                    24.11.1
webencodings                 0.5.1
websocket-client             1.8.0
websockets                   15.0.1
Werkzeug                     3.1.3
wheel                        0.45.1
wrapt                        1.14.2